setup: create basic communication to the companion app
This includes additional setup like the /common folder which for now
is only used for definitions to avoid possible errors while typing
hard coded strings, and to (in my eyes) better code readability.

The communication is based on simple strings. If complexity grows in
the future, this can be changed to json strings, but for simple messages
it is not needed.