logic: add main application logic
On startup (for now), the app sends a message to the companion to start
the wake-up interval timer. Once the timer is done, the companion sends
a message back to the app so that the app can check for the sleep state.

Finally, if the user is sleeping, we do nothing and the logic ends, but
if the user is not sleeping, we send a nudge and restart the timer.